refactor(frontend): move schemas to new subfolder; update imports
